python/requirements.txt
requests==2.22.0
python-socketio==5.6.0